tcc, core: Add isPrimary flag to ApplicationSpecificLB (#1365)
* primary lb changes

* tcc, core: Allow to have application across namespaces (#1292)

This is the summary of the proto change that I made:

namespaces - as scoping unit and not a hierarchy under Application. Not hierarchical because

Permission and access enforcement is at Application level and not at scoping unit

No metrics aggregation of any type and scoping unit

Resource (services) inside a scoping unit are unique across Application and not within scoping unit. Essentially a service with a same name can't exist in another scoping unit of the same application

Within an App and across scoping units, routing is enabled (has to be within the cluster)

Each scoping unit enables ingress into the applicaiton

Ingress routing is limited to services within the scope

An application can have scoping units without an Ingress

Scoping unit enables an application to be physically spread (across zones for e.g)

Am for now calling scoping units as namespaces. And it maps to physical namespace name in the cluster if that namespace exist in the cluster. We can improve the implementation to decouple this tie up later.

Implementation is getting a bit complicated. I haven't tested the changes fully, but to major extent is working ok. I need to add some more changes to the deployment yamls, regsource consumer, creating sidecars for cross namespace communication. Shall address them sometime today.

Mirrored from https://github.com/tetrateio/tetrate @ b6096f66968c5107909a00d2a544fb98ac66c51e