Operator cmn pkg (#3229)
This PR is part of splitting the PR #3202 into multiple PRs.

These are common packages

Compile, lint and test are disabled as the package refers to upstream changes that are yet to be merged. Except for changes mentioned below, all the structs, interfaces and functions are reused from upstream code.

helm package - a minor modification has to be done to prevent reference to upstream package vfs. vfs package has the compiled in charts. This naturally would be different for different operators. To make this package generic, a VFS interface is created and referenced.

name package - this is generic interface created to refer to the component names and related method for an operator. Upstream has used global package variable for this making all the places where name package was referred not re-usable.

helmreconciler package - Main change from upstream is to modify the RenderingInput (this naturally would refer to the particular operator spec). Embedding of struct used to reuse as much code as possible. Main code change necessitated to overcome upstream reference to name package and RenderingInput.

translate package - upstream struct embedded for code reuse. Changes necessitated due to reference of particular operator in upstream and reference to name package.

* common packages for operator implementation

* Added otel, prometheus and grafana the component and values-types protos

* update istio dep

* k8s go-client compatibility changes, istio api networking compatibiity
changes, lint

Mirrored from https://github.com/tetrateio/tetrate @ a443e2dae74fd39a761ef8357d3951ae4af43078